% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGO.r
\name{get_all_genes}
\alias{get_all_genes}
\title{get_all_genes}
\usage{
get_all_genes(ont = "BP")
}
\arguments{
\item{ont}{one of the 'BP', 'MF', or 'CC'}
}
\value{
a vector that contains all genes that have GO annotation
}
\description{
extracts all genes that have GO annotation
}
