% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGO.r
\name{get_all_GO}
\alias{get_all_GO}
\title{get_all_GO}
\usage{
get_all_GO(ont = "BP")
}
\arguments{
\item{ont}{one of the 'BP', 'MF', or 'CC'}
}
\value{
a vector that contains all GO terms of the specific 'ont'
}
\description{
extracts all the GO IDs that belongs to a specific GO branch
}
