% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGO-package.R
\docType{package}
\name{simpleGO-package}
\alias{simpleGO-package}
\title{simpleGO: Simple GO Enrichment Analysis}
\description{
Provides a simple implementation of GO ORA analysis for demonstration of R package development.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

}
\keyword{internal}
