% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGO.R
\name{simpleGO}
\alias{simpleGO}
\title{simpleGO}
\usage{
simpleGO(gene.list, background = NA, p = 0.05, excel.export = NA,
  use.synonims = FALSE)
}
\arguments{
\item{gene.list}{\code{character vector} with the gene set you want to enrich. Supports Human or Mouse, Gene Symbols or Ensembl.}

\item{background}{If you want to provide a custom background, again as \code{character vector}.}

\item{p}{Cutoff for the Benjamini corrected p-values.}

\item{excel.export}{A \code{character} with the file name in which you want to save the GO enrichments.}

\item{use.synonims}{If you are providing genes as Symbols, this will increase the number of recognized IDs. However, it can produce an increase in the number of your genes beacuse of older, redundant synonims shared between different genes.}
}
\value{
A \code{list} in which each element is a data.frame with the results of the enrichment for each GO functional cluster.
}
\description{
Simple GO enrichment with automatic clustering of the GO terms
}
\examples{
simpleGO(gene.list=c('Celf4','Eapp','Fzd8','Larp1'),p=0.01,excel.export="test.xlsx")

}
