% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGO.r
\name{simpleGO}
\alias{simpleGO}
\title{simpleGO}
\usage{
simpleGO(gene, GOID, background)
}
\arguments{
\item{gene}{a gene vector (e.g., DE genes)}

\item{GOID}{selected GO ID to test with}

\item{background}{background genes (e.g., all genes that have GO annotation)}
}
\value{
a data frame that contains a summary of the analysis
}
\description{
Provides a simple method for GO ORA analysis
}
\examples{
data(geneList, package = "DOSE")
de <- names(geneList)[1:100]
background <- get_all_genes('BP')
simpleGO(de, "GO:0007059", background)
}
